Created a Visual Studio solution.

+++ src/1.cc
@@ -35,7 +35,7 @@ void hexDecode(const char* input, uint8_t* out) {
 
   for (int i = 0; i < ((int)myString.size()) / 2; i++) {
     std::string subStr = myString.substr(i * 2, 2);
-    out[i] = strtoul(subStr.c_str(), 0, 16);
+    out[i] = (uint8_t) strtoul(subStr.c_str(), 0, 16);
   }
 }
